requirements.txt
google-generativeai
google-api-python-client
streamlit
gspread